+++ packages/plc-code/src/plc_code/formatcheck/__init__.py
@@ -8,5 +8,16 @@ from plc_code.formatcheck.checks import (
     check_xml,
     scan_block,
 )
+from plc_code.formatcheck.runner import Report, check_file, check_path
 
-__all__ = ["Finding", "BlockHeader", "check_bytes", "check_text", "check_xml", "scan_block"]
+__all__ = [
+    "Finding",
+    "BlockHeader",
+    "check_bytes",
+    "check_text",
+    "check_xml",
+    "scan_block",
+    "Report",
+    "check_file",
+    "check_path",
+]